Make limits 64 bits, add globallines limit to configurable limits
Make all limit counters 64 bits, in case someone really has a usage
for an insanely large program. The globallines limit was omitted, add
it to the list of configurable limits.

+++ asm/eval.c
@@ -72,7 +72,7 @@ static void *scpriv;
 static int *opflags;
 
 static struct eval_hints *hint;
-static int deadman;
+static int64_t deadman;
 
 
 /*
